Add inheritance for tests
