+++ src/test/java/Epics.java
@@ -6,22 +6,8 @@ import entities.Epic;
 import org.testng.Assert;
 import org.testng.annotations.*;
 
-public class Epics {
+public class Epics extends EpicBases{
 
-    static ApiRequest apiRequest = new ApiRequest();
-    public Epic epic;
-
-    @BeforeTest
-    public void setGeneralConfig() {
-        apiRequest.addHeader("X-TrackerToken", "599e3817e376dc26345552c4aa198143");
-        apiRequest.setBaseUri("https://www.pivotaltracker.com/services/v5");
-    }
-    @BeforeMethod(onlyForGroups = "getAEpic")
-    public void getAEpicConfig() {
-        apiRequest.setEndpoint("/projects/2505284/epics/{epic_id}");
-        apiRequest.setMethod(ApiMethod.GET);
-        apiRequest.addPathParam("epic_id", "4790780");
-    }
     @Test
     public void ItShouldGetAllEpicsOKStatusCode() {
         apiRequest.setEndpoint("/projects/2505284/epics");
@@ -97,23 +83,6 @@ public class Epics {
         int actual = apiResponse.getStatusCode();
         Assert.assertEquals(actual,expected);
     }
-    @AfterMethod(onlyForGroups = "CreateAEpic")
-    public void CleanCreatedEpic() {
-        apiRequest.setEndpoint("/projects/2505284/epics/{epic_id}");
-        apiRequest.setBody("");
-        apiRequest.addPathParam("epic_id", epic.getId());
-        apiRequest.setMethod(ApiMethod.DELETE);
-        ApiResponse apiResponse = ApiManager.execute(apiRequest);
-    }
-
-    @BeforeMethod(onlyForGroups = "DeleteAEpic")
-    public void deleteEpicsConfig() {
-        apiRequest.setEndpoint("/projects/2505284/epics");
-        apiRequest.setBody("{\"name\":\"ThisEpicWillBeDeleted\"}");
-        apiRequest.setMethod(ApiMethod.POST);
-        ApiResponse apiResponse = ApiManager.executeWithBody(apiRequest);
-        epic = apiResponse.getBody(Epic.class);
-    }
 
     @Test(groups = "DeleteAEpic")
     public void deleteAEpic(){
@@ -126,11 +95,7 @@ public class Epics {
         int actual = apiResponse.getStatusCode();
         Assert.assertEquals(actual,expected);
     }
-    @AfterMethod
-    public void CleanObjects(){
-        epic = new Epic();
-        apiRequest.clearPathParms();
-    }
+
     @Test
     public void ShouldReturnBadRequestForIncorrectEpicID(){
         apiRequest.setEndpoint("/projects/2505284/epics/{epic_id}");
