+++ src/test/java/WorkSpacesTest.java
@@ -6,22 +6,8 @@ import entities.WorkSpaces;
 import org.testng.Assert;
 import org.testng.annotations.*;
 
-public class WorkSpacesTest {
+public class WorkSpacesTest extends WorkSpacesBases {
 
-    static ApiRequest apiRequest = new ApiRequest();
-    public WorkSpaces workSpaces;
-
-    @BeforeTest
-    public void setGeneralConfig() {
-        apiRequest.addHeader("X-TrackerToken", "599e3817e376dc26345552c4aa198143");
-        apiRequest.setBaseUri("https://www.pivotaltracker.com/services/v5");
-    }
-    @BeforeMethod(onlyForGroups = "getAWorkSpace")
-    public void getAProjectsConfig() {
-        apiRequest.setEndpoint("/my/workspaces/{workspace_id}");
-        apiRequest.setMethod(ApiMethod.GET);
-        apiRequest.addPathParam("workspace_id", "876161");
-    }
     @Test
     public void ItShouldGetAllWorkSpacesOKStatusCode() {
         apiRequest.setEndpoint("/my/workspaces");
@@ -76,14 +62,7 @@ public class WorkSpacesTest {
         String actual = workSpaces.getName();
         Assert.assertEquals(actual,expected);
     }
-    @AfterMethod(onlyForGroups = "CreateAWorkSpace")
-    public void CleanCreatedWorkSpace() {
-        apiRequest.setEndpoint("/my/workspaces/{workspace_id}");
-        apiRequest.setBody("");
-        apiRequest.addPathParam("workspace_id",workSpaces.getId());
-        apiRequest.setMethod(ApiMethod.DELETE);
-        ApiResponse apiResponse = ApiManager.execute(apiRequest);
-    }
+
     @Test
     public void shouldReturnBadRequestForInvalidWorkSpaceBody(){
         apiRequest.setEndpoint("/my/workspaces");
@@ -104,15 +83,6 @@ public class WorkSpacesTest {
         int actual = apiResponse.getStatusCode();
         Assert.assertEquals(actual,expected);
     }
-    @BeforeMethod(onlyForGroups = "DeleteAWorkSpace")
-    public void deleteProjectsConfig() {
-        apiRequest.setEndpoint("/my/workspaces");
-        apiRequest.setBody("{\"name\":\"WorkSpaceToBeDeleted\",\"project_ids\":[2505284]}");
-        apiRequest.setMethod(ApiMethod.POST);
-        ApiResponse apiResponse = ApiManager.executeWithBody(apiRequest);
-        workSpaces = apiResponse.getBody(WorkSpaces.class);
-    }
-
     @Test(groups = "DeleteAWorkSpace")
     public void deleteAProject(){
         apiRequest.setEndpoint("/my/workspaces/{workspace_id}");
@@ -124,11 +94,7 @@ public class WorkSpacesTest {
         int actual = apiResponse.getStatusCode();
         Assert.assertEquals(actual,expected);
     }
-    @AfterMethod
-    public void CleanObjects(){
-        workSpaces = new WorkSpaces();
-        apiRequest.clearPathParms();
-    }
+
     @Test
     public void ShouldReturnNotFoundForIncorrectEpicID(){
         apiRequest.setEndpoint("/my/workspaces/{workspace_id}");
